cleaned dx12 mesh manager

+++ SirEngineThe3rdLib/src/platform/windows/graphics/dx12/descriptorHeap.h
@@ -70,7 +70,7 @@ class SIR_ENGINE_API DescriptorHeap {
     int idx = findCPUDescriptorIndexFromHandle(handles.cpuHandle);
     // freeing is just a matter of freeing up the index
     // then it will get overwritten
-    assert(idx < m_freeList.size());
+    assert(idx < static_cast<int>(m_freeList.size()));
     assert(idx >= 0);
     m_freeList[m_freeListIdx++] = idx;
   }
